+++ CHANGES.md
@@ -9,6 +9,9 @@ Version 1.2.9
 
 To be released.
 
+ -  Fixed idempotence check in inbox listeners to ensure activities for
+    different origins are processed correctly.
+
 
 Version 1.2.8
 -------------
@@ -216,6 +219,15 @@ Released on October 31, 2024.
 [#118]: https://github.com/dahlia/fedify/issues/118
 
 
+Version 1.1.9
+-------------
+
+Released on December 11, 2024.
+
+ -  Fixed idempotence check in inbox listeners to ensure activities for
+    different origins are processed correctly.
+
+
 Version 1.1.8
 -------------
 
@@ -301,7 +313,7 @@ Released on November 19, 2024.
 Version 1.1.5
 -------------
 
-Released on December 14, 2024.
+Released on November 14, 2024.
 
  -  Suppressed a `TypeError` with a message <q>unusable</q> due to Node.js's
     mysterious behavior.  [[#159]]
@@ -463,10 +475,19 @@ Released on October 20, 2024.
 [#150]: https://github.com/dahlia/fedify/issues/150
 
 
+Version 1.0.12
+--------------
+
+Released on December 11, 2024.
+
+ -  Fixed idempotence check in inbox listeners to ensure activities for
+    different origins are processed correctly.
+
+
 Version 1.0.11
 --------------
 
-Released on December 22, 2024.
+Released on November 22, 2024.
 
  -  Fixed a bug where `lookupWebFinger()` function had thrown a `TypeError`
     when the *.well-known/webfinger* redirects to a relative URI.  [[#166]]
@@ -475,7 +496,7 @@ Released on December 22, 2024.
 Version 1.0.10
 --------------
 
-Released on December 19, 2024.
+Released on November 19, 2024.
 
  -  Fix a bug where `Actor`'s `inbox` and `outbox` properties had not been
     able to be set to an `OrderedCollectionPage` object, even though it is
@@ -532,7 +553,7 @@ Released on December 19, 2024.
 Version 1.0.9
 -------------
 
-Released on December 14, 2024.
+Released on November 14, 2024.
 
  -  Suppressed a `TypeError` with a message <q>unusable</q> due to Node.js's
     mysterious behavior.  [[#159]]